requirements.txt
flask
pytesseract
pillow
opencv-python-headless
numpy
